refactor: 패키지 분리 CarResponse, TrackRequest, TrackResponse를 controller/dto로 이동
