+++ src/main/java/racingcar/controller/RacingConsoleController.java
@@ -1,5 +1,6 @@
 package racingcar.controller;
 
+import racingcar.controller.dto.TrackRequest;
 import racingcar.exception.CustomException;
 import racingcar.model.car.Cars;
 import racingcar.model.car.strategy.MovingStrategy;
