+++ src/main/java/racingcar/controller/RacingWebController.java
@@ -5,6 +5,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import racingcar.controller.dto.TrackRequest;
+import racingcar.controller.dto.TrackResponse;
 import racingcar.service.RacingService;
 
 import java.util.List;
