+++ src/main/java/racingcar/controller/dto/TrackRequest.java
@@ -1,4 +1,4 @@
-package racingcar.controller;
+package racingcar.controller.dto;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
