+++ src/main/java/racingcar/controller/dto/TrackResponse.java
@@ -1,4 +1,4 @@
-package racingcar.controller;
+package racingcar.controller.dto;
 
 import java.util.List;
